Final Push
I wasn't able to get the button navigation from viewSavedQuotes to Home and InsertQuotes to Home to work, there is a weird UI object that only functions correct in mainWindow class so and I couldnt figure out how to properly access it in the InsertQuote/ViewSavedQuotes class. Instead, I touched up the UI so none of the text is cutt off or looking too weird, hopefully this is good enough.